refactor(projects): isolate locked policy repository
